+++ src/Entity/Archive.php
@@ -6,6 +6,8 @@ use App\Entity\Traits\SmallImage1Trait;
 use App\Entity\Traits\SmallImage2Trait;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\HttpFoundation\File\File;
+use Symfony\Component\Validator\Constraints as Assert;
 use Vich\UploaderBundle\Mapping\Annotation as Vich;
 
 /**
@@ -23,6 +25,20 @@ class Archive extends AbstractBase
      * @ORM\Column(type="integer", unique=true)
      */
     private int $year;
+
+    /**
+     * @Assert\File(maxSize="10M", mimeTypes={"image/png", "image/jpeg", "image/pjpeg"})
+     * @Assert\Image(allowSquare=true, allowLandscape=false, allowPortrait=false, minWidth=400, minHeight=400)
+     * @Vich\UploadableField(mapping="image", fileNameProperty="smallImage1FileName")
+     */
+    private ?File $smallImage1File = null;
+
+    /**
+     * @Assert\File(maxSize="10M", mimeTypes={"image/png", "image/jpeg", "image/pjpeg"})
+     * @Assert\Image(allowSquare=true, allowLandscape=false, allowPortrait=false, minWidth=400, minHeight=400)
+     * @Vich\UploadableField(mapping="image", fileNameProperty="smallImage2FileName")
+     */
+    private ?File $smallImage2File = null;
 
     public function getYear(): int
     {