+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
  * add missing info cards in default page view
  * add missing brand new favicon black
  * hide unnecessary ConfigFooterInformation admin
+ * add Archive admin help image to only allow squared aspects
 
 ##### Version 5.08.00 (2021-11-25)
  * fix broken testing suite